tke/requirements.txt
Django==1.8.4
antlr4-python2-runtime==4.5.2
appnope==0.1.0
decorator==4.0.2
django-extensions==1.5.5
django-json-field==0.5.7
#gnureadline==6.3.3
ipdb==0.8.1
ipython==4.0.0
ipython-genutils==0.1.0
jsonfield==1.0.3
lxml==3.4.4
path.py==7.6.1
pexpect==3.3
pickleshare==0.5
python-dateutil==2.4.2
simplegeneric==0.8.1
six==1.9.0
traitlets==4.0.0
wsgiref==0.1.2
